规范修改

+++ 8301-stock/src/main/java/com/mmwwtt/stock/entity/strategy/BaseStrategy.java
@@ -31,7 +31,7 @@ public class BaseStrategy {
      */
     @EqualsAndHashCode.Include
     @TableId(type = IdType.AUTO)
-    protected Long strategyId;
+    protected Integer strategyId;
 
     /**
      * 策略编码
